+++ plugin/src/net/groboclown/idea/p4ic/ui/config/props/RelativeConfigPartPanel.java
@@ -14,6 +14,7 @@
 
 package net.groboclown.idea.p4ic.ui.config.props;
 
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
@@ -39,10 +40,20 @@ public class RelativeConfigPartPanel
 
     RelativeConfigPartPanel(@NotNull Project project, @NotNull final RelativeConfigCompositePart part) {
         super(project, part);
-        if (part.getName() != null) {
-            part.setName(DEFAULT_FILE_NAME);
+        final String name = part.getName();
+        if (name == null) {
+            nameField.setText(DEFAULT_FILE_NAME);
+            // Setting the name will trigger a reload, so do that in the background
+            // See #139
+            ApplicationManager.getApplication().executeOnPooledThread(new Runnable() {
+                @Override
+                public void run() {
+                    part.setName(DEFAULT_FILE_NAME);
+                }
+            });
+        } else {
+            nameField.setText(name);
         }
-        nameField.setText(part.getName());
     }
 
     @Nls